fix(pipeline): close delta round 2 (retry budget, terminal matrix, fence)
- send_with_retry/send_stream_with_retry now OWN the per-request
  timeout as a total budget (retries get only the remainder — a
  sequence cannot outlive its bound); 14 call sites dropped their
  builder .timeout; the streaming twin's duplicated stream_deadline
  argument collapsed. QUALITY_RUN_FIXED_SECS 4200 is now correct
  rather than accidental; client margin 60->960s (worst in-flight
  overshoot is stream_deadline(max)); argued deviation: MAX_ATTEMPTS
  no longer bounds anything, so no test couples to it — the budget is
  the bound and IS mutation-pinned
- terminal_state takes persisted: a boundary RunTimeout with a saved
  document is needsReview/completed, and the job event moved with it
  (job_complete whenever status != failed)
- complete_json takes a required deadline guard composed into charge —
  the re-ask cannot fire past deadline; single source for the
  run-timeout message
- fence neutralizer writes the matched run BACK instead of dropping it
  (the executed proof showed 3 posting lines deleted); the reviewer's
  prescribed {0,200} bound was probed and REJECTED — padded attributes
  survive it; pattern stays unbounded+newline-tolerant, hostile set
  96 tests green; agent/tools.rs split at R8 (tests -> tools/test.rs)
- emphasisDropped count on the strategy artifact; repair timing test
  margins one-sided; is_verbatim cfg(test); ensure_latest_run message
  honest; TS lock notes the Rust twin is the live guard
